refactor(#130): 인증 관련 컴포넌트 구조 개선
