+++ apps/web/src/components/signup/index.ts
@@ -1,3 +1,3 @@
-export { default as SignUpContent } from './SignUpContent/SignUpContent';
+export { default as SignUpContent } from '@/clients/SignUpContent/SignUpContent';
 export { default as SecondaryAction } from './SecondaryAction/SecondaryAction';
 export { default as VerificationInput } from './VerificationInput/VerificationInput';
