+++ apps/web/src/hooks/useSignUpHandlers.ts
@@ -2,7 +2,7 @@ import { useRegister } from '@/hooks/api/useAuth';
 import { useRouter } from 'next/navigation';
 import type React from 'react';
 import { useState } from 'react';
-import type { SignUpData, SignUpStep } from '../types';
+import type { SignUpData, SignUpStep } from '../types/types';
 import { useSignUpValidation } from './useSignUpValidation';
 
 export const useSignUpHandlers = () => {
